+++ MicroBlink.framework/Headers/PPAusIDFrontRecognizerResult.h
@@ -17,30 +17,30 @@ PP_CLASS_AVAILABLE_IOS(6.0) @interface PPAusIDFrontRecognizerResult : PPBlinkOcr
  * Returns the first name of the Austrian ID owner.
  * @return first name of the Austrian ID owner.
  */
-- (NSString *)firstName;
+- (NSString * _Nullable)firstName;
 
 /**
  * Returns the last name of the Austrian ID owner.
  * @return last name of the Austrian ID owner.
  */
-- (NSString *)lastName;
+- (NSString * _Nullable)lastName;
 
 /**
  * Returns the document number of the Austrian ID.
  * @return document number of the Austrian ID.
  */
-- (NSString *)documentNumber;
+- (NSString * _Nullable)documentNumber;
 
 /**
  * Returns the date of birth of the Austrian ID owner.
  * @return date of birth of the Austrian ID owner.
  */
-- (NSString *)dateOfBirth;
+- (NSString * _Nullable)dateOfBirth;
 
 /**
  * Returns the sex of the Austrian ID owner.
  * @return sex of the Austrian ID owner.
  */
-- (NSString *)sex;
+- (NSString * _Nullable)sex;
 
 @end