+++ MicroBlink.framework/Headers/PPBlinkInputRecognizers.h
@@ -1,0 +1,23 @@
+//
+//  PPBlinkInputRecognizers.h
+//  BlinkOcrFramework
+//
+//  Created by Dino on 27/09/16.
+//  Copyright © 2016 MicroBlink Ltd. All rights reserved.
+//
+
+#ifndef PPBlinkInputRecognizers_h
+#define PPBlinkInputRecognizers_h
+
+
+#import "PPBlinkBarcodeRecognizers.h"
+
+/** OCR */
+#import "PPBlinkOcrRecognizerSettings.h"
+#import "PPBlinkOcrRecognizerResult.h"
+
+/** Detector */
+#import "PPDetectorRecognizerSettings.h"
+#import "PPDetectorRecognizerResult.h"
+
+#endif /* PPBlinkInputRecognizers_h */