+++ MicroBlink.framework/Headers/PPDetectorRecognizerResult.h
@@ -15,8 +15,6 @@
 #import "PPMultiDetectorResult.h"
 #import "PPPointsDetectorResult.h"
 
-NS_ASSUME_NONNULL_BEGIN
-
 /**
  * Recognizer wrapper for PPDetectorResult
  */
@@ -28,8 +26,6 @@ PP_CLASS_AVAILABLE_IOS(6.0) @interface PPDetectorRecognizerResult : PPRecognizer
  *
  *  @return detectorResult
  */
-- (PPDetectorResult *)detectorResult;
+- (PPDetectorResult * _Nullable)detectorResult;
 
 @end
-
-NS_ASSUME_NONNULL_END