+++ MicroBlink.framework/Headers/PPSerbianIdBackRecognizerSettings.h
@@ -1,0 +1,33 @@
+//
+//  PPSerbianIdBackRecognizerSettings.h
+//  BlinkIdFramework
+//
+//  Created by DoDo on 09/11/2016.
+//  Copyright © 2016 MicroBlink Ltd. All rights reserved.
+//
+
+#import "PPRecognizerSettings.h"
+
+/**
+ * Settings class for configuring Serbian Back ID Recognizer
+ *
+ * Serbian Back ID recognizer is used for scanning back side of Serbian ID.
+ */
+PP_CLASS_AVAILABLE_IOS(6.0) @interface PPSerbianIdBackRecognizerSettings : PPRecognizerSettings
+
+/**
+ * Name of the image sent to didOutputMetadata method of scanDelegate object that contains full document.
+ * This image will be sent to scan delegate during recognition process if displaying of full document image
+ * is enabled via displayFullDocumentImage property and receiving of dewarpedImage in MetadataSettings is enabled.
+ */
+@property (nonatomic, readonly) NSString* FULL_DOCUMENT_IMAGE;
+
+/**
+ * Sets whether full document image of ID card should be sent to didOutputMetadata method of scanDelegate object.
+ * If you want to recieve this image, be sure to enable dewarpedImage in MetadataSettings.
+ *
+ * Default: NO
+ */
+@property (nonatomic, assign) BOOL displayFullDocumentImage;
+
+@end