+++ MicroBlink.framework/Headers/PPSlovakIdFrontRecognizerResult.h
@@ -8,8 +8,6 @@
 
 #import "PPBlinkOcrRecognizerResult.h"
 
-NS_ASSUME_NONNULL_BEGIN
-
 /**
  * Class representing values obtained when scanning front side of Slovak ID
  */
@@ -19,62 +17,60 @@ PP_CLASS_AVAILABLE_IOS(6.0) @interface PPSlovakIdFrontRecognizerResult : PPBlink
  * Returns the first name of the Slovakian ID owner.
  * @return the first name of the Slovakian ID owner.
  */
-- (NSString *)firstName;
+- (NSString * _Nullable)firstName;
 
 /**
  * Returns the last name of the Slovakian ID owner.
  * @return the last name of the Slovakian ID owner.
  */
-- (NSString *)lastName;
+- (NSString * _Nullable)lastName;
 
 /**
  * Returns the sex of the Slovakian ID owner.
  * @return the sex of the Slovakian ID owner.
  */
-- (NSString *)sex;
+- (NSString * _Nullable)sex;
 
 /**
  * Returns the nationality of the Slovakian ID owner.
  * @return the nationality of the Slovakian ID owner.
  */
-- (NSString *)nationality;
+- (NSString * _Nullable)nationality;
 
 /**
  * Returns the date of birth of the Slovakian ID owner.
  * @return the date of birth of the Slovakian ID owner.
  */
-- (NSString *)dateOfBirth;
+- (NSString * _Nullable)dateOfBirth;
 
 /**
  * Returns the date of expiry of the Slovakian ID.
  * @return the date of expiry of the Slovakian ID.
  */
-- (NSString *)dateOfExpiry;
+- (NSString * _Nullable)dateOfExpiry;
 
 /**
  * Returns the date of issue of the Slovakian ID.
  * @return the date of issue of the Slovakian ID.
  */
-- (NSString *)dateOfIssue;
+- (NSString * _Nullable)dateOfIssue;
 
 /**
  * Returns the issuing authority of the ID.
  * @return the issuing authority of the ID.
  */
-- (NSString *)issuedBy;
+- (NSString * _Nullable)issuedBy;
 
 /**
  * Returns the personal number of the Slovakian ID owner.
  * @return the personal number of the Slovakian ID owner.
  */
-- (NSString *)personalNumber;
+- (NSString * _Nullable)personalNumber;
 
 /**
  * Returns the document number of the Slovakian ID.
  * @return the document number of the Slovakian ID.
  */
-- (NSString *)documentNumber;
+- (NSString * _Nullable)documentNumber;
 
 @end
-
-NS_ASSUME_NONNULL_END