+++ MicroBlink.framework/Headers/PPSlovenianIdFrontRecognizerResult.h
@@ -8,8 +8,6 @@
 
 #import "PPBlinkOcrRecognizerResult.h"
 
-NS_ASSUME_NONNULL_BEGIN
-
 /**
  * Class representing values obtained when scanning front side side of Slovenian ID
  */
@@ -19,38 +17,36 @@ PP_CLASS_AVAILABLE_IOS(6.0) @interface PPSlovenianIdFrontRecognizerResult : PPBl
  * Returns the first name of the Slovenian ID owner.
  * @return the first name of the Slovenian ID owner.
  */
-- (NSString *)firstName;
+- (NSString * _Nullable)firstName;
 
 /**
  * Returns the last name of the Slovenian ID owner.
  * @return the last name of the Slovenian ID owner.
  */
-- (NSString *)lastName;
+- (NSString * _Nullable)lastName;
 
 /**
  * Returns the sex of the Slovenian ID owner.
  * @return the sex of the Slovenian ID owner.
  */
-- (NSString *)sex;
+- (NSString * _Nullable)sex;
 
 /**
  * Returns the nationality of the Slovenian ID owner.
  * @return the nationality of the Slovenian ID owner.
  */
-- (NSString *)nationality;
+- (NSString * _Nullable)nationality;
 
 /**
  * Returns the date of birth of the Slovenian ID owner.
  * @return the date of birth of the Slovenian ID owner.
  */
-- (NSString *)dateOfBirth;
+- (NSString * _Nullable)dateOfBirth;
 
 /**
  * Returns the date of expiry of the Slovenian ID owner.
  * @return the date of expiry of the Slovenian ID owner.
  */
-- (NSString *)dateOfExpiry;
+- (NSString * _Nullable)dateOfExpiry;
 
 @end
-
-NS_ASSUME_NONNULL_END